🔄 test classes now run with actual React DOM
