+++ Framework/Testing/ShellTestRunner.js
@@ -1,42 +1,54 @@
-class ShellTestRunner {
-  constructor() {
-    this.className = 'ShellClass';
-  }
+import React, { Node, useCallback, memo, useEffect, useState } from 'react';
+import { View, LogBox, Platform, StatusBar } from 'react-native';
+
+/**
+ * Run unit/integration tests for a single class with an actual React Native DOM.
+ * 
+ * @param {Object} props - Component props.
+ * @param {Function} props.logClassTestResult - Logs test results given a list of results.
+ */
+const ShellTestRunnerComp = ({logClassTestResult}) => {
+  const className = "Sample";
+
+  useEffect(() => {
+    runTests();
+  }, []);
 
   /**
    * Runs all tests for this module synchronously.
-   * @returns {Array} An array of test results.
    */
-  runTests() {
+  async function runTests() {
     const results = [];
 
-    results.push({ test: 'exampleTest', status: this.exampleTest() });
-    results.push({ test: 'exampleTest2', status: this.exampleTest2() });
+    results.push({ test: 'exampleTest', status: await exampleTest() });
+    results.push({ test: 'exampleTest2', status: await exampleTest2() });
 
-    return results;
+    logClassTestResult(className, results);
   }
 
   /**
    * Example test case.
    * 
-   * @returns {Boolean} Test result status.
+   * @returns {Promise<boolean>} A promise that resolves to the test result status.
    */
-  exampleTest() {
+  async function exampleTest() {
     let status = true;
 
     return status;
   }
 
-    /**
-   * Example test case.
+  /**
+   * Example test case 2.
    * 
-   * @returns {Boolean} Test result status.
+   * @returns {Promise<boolean>} A promise that resolves to the test result status.
    */
-    exampleTest2() {
-      let status = false;
-  
-      return status;
-    }
+  async function exampleTest2() {
+    let status = false;
+
+    return status;
+  }
+
+  return null;
 }
 
-export default ShellTestRunner;
+export default memo(ShellTestRunnerComp);
